fixed the scrolling issue
Closes #11
