+++ App.jsx
@@ -1,35 +1,10 @@
 import React, { useState } from 'react';
 import { StyleSheet, View } from 'react-native';
-import {
-  extendTheme,
-  NativeBaseProvider,
-  Text,
-  Box,
-  Stack,
-  StatusBar,
-} from 'native-base';
+import { extendTheme, NativeBaseProvider } from 'native-base';
 import AppLoading from 'expo-app-loading';
 import * as Font from 'expo-font';
-import LoginScreen from './Components/Logins/LoginScreen';
-import RegisterScreen from './Components/Logins/RegisterScreen';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { createStackNavigator } from '@react-navigation/stack';
-import { NavigationContainer } from '@react-navigation/native';
-import Setting from './Components/Screens/Settings';
-import Home from './Components/Screens/Home';
-import Donations from './Components/Screens/Donations';
-import { Ionicons } from '@expo/vector-icons';
-import { CommonStrings } from './Styles/CommonStrings';
-import HomeDetails from './Components/Detail-Screens/HomeDetails';
-import DonationDetails from './Components/Detail-Screens/DonationDetails';
-import { Colors } from './Styles/Colors';
-import { enableScreens } from 'react-native-screens';
-import GoogleAuthLogin from './Components/Logins/GoogleAuthLogin';
-import FilterDetail from './Components/Detail-Screens/FilterDetail';
-import { TransitionPresets } from '@react-navigation/stack';
-import ToTheReader from './Components/Detail-Screens/ToTheReader';
-
-enableScreens();
+import { AuthProvider } from './hooks/auth';
+import MainScreen from './Components/MainScreen/MainScreen';
 
 //fonts
 const theme = extendTheme({
@@ -38,102 +13,6 @@ const theme = extendTheme({
     heading: 'Bold',
   },
 });
-
-const Tab = createBottomTabNavigator();
-const HomeStack = createStackNavigator();
-
-/* Bottom tab configuration */
-
-const NavTab = () => {
-  return (
-    <Tab.Navigator
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ focused, color, size }) => {
-          let iconName;
-          let accentColor;
-          if (route.name === 'Adopt') {
-            iconName = focused ? 'paw' : 'paw-outline';
-          } else if (route.name === 'Donations') {
-            iconName = focused ? 'ios-gift' : 'ios-gift-outline';
-          } else if (route.name === 'Settings') {
-            iconName = focused ? 'ios-settings' : 'ios-settings-outline';
-          }
-
-          // You can return any component that you like here!
-          return <Ionicons name={iconName} size={22} color={color} />;
-        },
-        tabBarActiveTintColor: Colors.metalGray,
-        tabBarStyle: {
-          color: Colors.metalGray,
-          fontFamily: 'Medium',
-        },
-      })}
-    >
-      <Tab.Screen
-        options={{ headerShown: false }}
-        name={CommonStrings.homes}
-        component={Home}
-      />
-      <Tab.Screen
-        options={{ headerShown: false }}
-        name={CommonStrings.donation}
-        component={Donations}
-      />
-      <Tab.Screen
-        options={{ headerShown: false }}
-        name={CommonStrings.setting}
-        component={Setting}
-      />
-    </Tab.Navigator>
-  );
-};
-
-const BottomStack = createStackNavigator();
-
-/** This will help keep the stack navigation in the bottom */
-
-const Navigation = () => {
-  return (
-    <NavigationContainer>
-      <BottomStack.Navigator
-        screenOptions={{
-          headerShown: false,
-        }}
-      >
-        <BottomStack.Group>
-          <BottomStack.Screen
-            name={CommonStrings.home}
-            children={() => <NavTab />}
-          />
-          <BottomStack.Screen name="HomeDetailScreen" component={HomeDetails} />
-          <BottomStack.Screen
-            name="DonationDetailScreen"
-            component={DonationDetails}
-            options={{
-              ...TransitionPresets.ModalSlideFromBottomIOS,
-            }}
-          />
-        </BottomStack.Group>
-        <BottomStack.Group screenOptions={{ presentation: 'modal' }}>
-          <BottomStack.Screen
-            name="FilterDetailScreen"
-            component={FilterDetail}
-            options={{
-              ...TransitionPresets.ModalPresentationIOS,
-            }}
-          />
-          <BottomStack.Screen
-            name="ToTheReaderDetail"
-            component={ToTheReader}
-            options={{
-              ...TransitionPresets.ModalPresentationIOS,
-            }}
-          />
-        </BottomStack.Group>
-      </BottomStack.Navigator>
-    </NavigationContainer>
-  );
-};
 
 /***
  * Async font loading. Please look at the expo for examples
@@ -150,41 +29,14 @@ const getFonts = () => {
 export default function App() {
   const [fontLoaded, setFontLoaded] = useState(false);
 
-  const [user, setUser] = useState(1);
-
   if (fontLoaded) {
-    if (user) {
-      return (
-        <NativeBaseProvider theme={theme}>
-          <Navigation />
-        </NativeBaseProvider>
-      );
-    } else {
-      return (
-        <NativeBaseProvider theme={theme}>
-          <NavigationContainer>
-            <HomeStack.Navigator>
-              {user ? (
-                <Navigation />
-              ) : (
-                <>
-                  <HomeStack.Screen
-                    options={{ headerShown: false }}
-                    name="Login"
-                    component={GoogleAuthLogin}
-                  />
-                  <HomeStack.Screen
-                    options={{ headerShown: false }}
-                    name="Register"
-                    component={RegisterScreen}
-                  />
-                </>
-              )}
-            </HomeStack.Navigator>
-          </NavigationContainer>
-        </NativeBaseProvider>
-      );
-    }
+    return (
+      <NativeBaseProvider theme={theme}>
+        <AuthProvider>
+          <MainScreen />
+        </AuthProvider>
+      </NativeBaseProvider>
+    );
   } else {
     return (
       <AppLoading
